Implemented Crafting Ability & Item Drop
Added the ability to discard items from the inventory by dragging them out of the menus. Added a pop-up confirmation window before it fully executes.

Added the ability to craft items together to make a new one and destroy the original two. Currently only works when combining Trash & Trash but will duplicate the code to allow for the other combinations next. The crafted item is not added to the crafted item inventory as an actualy item yet. That will also be the next step.