CRAFTING FINALLY WORKS
Crafting system is fully implemented. Any two items can be dragged into the crafting window from the inventory panel and crafted into a new one. Those items are destroyed and the new item momentarily appears in the crafting window before being placed in the crafted items panel. Code is in place to prohibit players from crafting when the crafted items panel is full. Code is in place that alerts players if they do not have enough items to craft when the button is clicked. Items can be discarded by dragging them out of the inventory/crafting panels, prompting a pop up to ask the player to confirm their discard.

All that is left to do is add the ability to use an item from the inventory screen. I think I'm going to create a new popup window that asks if the player wants to use the item or discard it, and then activate the confirm discard window if they choose discard.