Numerous Crafting System Touchups
Crafting System is fully functional. There are a few minor bugs that can either be worked out by the dev team once class starts, or be embraced as one of a kind features but the system fully works.

All that's left to be done is implement the ability to use an item from inventory and have it change aspects of the game. Pseudo code/starting code is in place in the EquipableItem.cs script. Just need to figure out how to call the functions when an item is used, and then implement a timer within the functions so the effect doesn't last forever.