Implemented Beginnings Of Environmental Hiding
Created a script to attach to 3D objects that players can hide inside of. Script prevents enemies from seeing the player while they are inside the object. The object has a mesh renderer and a trigger collider to allow that player to walk into the object. Once inside the object, the screen darkens (to indicate that they are in hiding) and a timer bar appears and shrinks to indicate how long they can stay hidden. Once the timer is up the screen brightens back to normal. However, the object does not disappear or kick the player out. Still need to implement limit on useage of hiding spots and variables to activate the ability to use them.