Initial Enemy Import
Created a 3D model for Manager Mark and applied four different animations to the model. Imported the model and the animations. Setup an animator controller to control which animations play for the enemy and when. Edited the EnemySight script to call for the different animations when the State changes. Applied the EnemySight to the new imported enemy model and rebaked the NavMesh for the level.