Implemented Enemy Audio Successfully
Enemy has an audio source that switches the sound being played in accordance with the state that the enemy is currently in. When switching to the chase state, script compares whether state was already state before starting chase audio so it isn't constantly trying to play the audio every frane (resulting in silence). All audio is stopped when the enemy catches the player.

Also fixed some bugs in enemy AI where investigate state only worked once.